{{ cookiecutter.project_slug }}/requirements.txt
appdirs==1.4.4
astroid==2.7.2
attrs==21.2.0
backcall==0.2.0
black==21.7b0
click==8.0.1
colorama==0.4.4
coverage==5.5
decorator==5.0.9
flasgger==0.9.5
Flask==2.0.1
gunicorn==20.1.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.2.0
pluggy==0.13.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pylint==2.10.2
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-sugar==0.9.4
python-decouple==3.4
python-json-logger==2.0.2
PyYAML==5.4.1
regex==2021.8.28
six==1.16.0
termcolor==1.1.0
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
